Replace cecum1 with cecum2
