+++ src/scaffoldmaker/scaffolds.py
@@ -19,7 +19,6 @@ from scaffoldmaker.meshtypes.meshtype_3d_boxhole1 import MeshType_3d_boxhole1
 from scaffoldmaker.meshtypes.meshtype_3d_boxnetwork1 import MeshType_3d_boxnetwork1
 from scaffoldmaker.meshtypes.meshtype_3d_brainstem import MeshType_3d_brainstem1
 from scaffoldmaker.meshtypes.meshtype_3d_cecum1 import MeshType_3d_cecum1
-from scaffoldmaker.meshtypes.meshtype_3d_cecum2 import MeshType_3d_cecum2
 from scaffoldmaker.meshtypes.meshtype_3d_colon1 import MeshType_3d_colon1
 from scaffoldmaker.meshtypes.meshtype_3d_colonsegment1 import MeshType_3d_colonsegment1
 from scaffoldmaker.meshtypes.meshtype_3d_esophagus1 import MeshType_3d_esophagus1
@@ -75,7 +74,6 @@ class Scaffolds(object):
             MeshType_3d_boxnetwork1,
             MeshType_3d_brainstem1,
             MeshType_3d_cecum1,
-            MeshType_3d_cecum2,
             MeshType_3d_colon1,
             MeshType_3d_colonsegment1,
             MeshType_3d_esophagus1,
